% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_std.R
\name{baseline_std}
\alias{baseline_std}
\title{This function takes data from MS lesions and standardizes the intensities
based on the pre-incidence or baseline values.}
\usage{
baseline_std(xdat, method = "stabilized")
}
\arguments{
\item{xdat}{Data must be set up in long format with the following variables:
lid = lesion id (factor or numeric)
vid = voxel id (factor or numeric)
vtime = time variable, can be visit number or days from incidence (numeric), where 0 = time of lesion incidence.
y = MRI instensity}

\item{method}{There are three standaridization methods available: Voxel level, lesion level, and stabilized versions. method = "stabilized", "voxel", or "lesion"}
}
\value{
Output is original data (df) with the following new columns/variables:
y_std = standardized intensities at each time point
base_sd = baseline sd used to calculate standardized intensities (can be unique for each vid)
base_mean = baseline mean intensity for pre-incidence intensities for "current" vid
}
\description{
This function takes data from MS lesions and standardizes the intensities
based on the pre-incidence or baseline values.
}
