% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{dlm}
\alias{dlm}
\title{Fit dynamic prediction model}
\usage{
dlm(dat, htimes, ptimes, nfit = NULL, tx = 1)
}
\arguments{
\item{dat}{has three columns vid, itime, and y_i, same as data from output of interp_times}

\item{htimes}{is a vector of historical times included as covariates in model}

\item{ptimes}{is vector of prediction times in the future}

\item{nfit}{is the number of voxels to predict for model, when NULL runs all}

\item{tx}{is used for parallelization, it is the tx^th iteration of the code that is run
tx along with nfit determines which voxels are predicted from model (line 27)}
}
\value{
list with two objects dlm_model (a list with DLM objects from each prediction time point)
#' and predictions, a matrix of predicted intensities for the prediction times
}
\description{
Fit dynamic prediction model
}
