% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_pred.R
\name{dlm_pred}
\alias{dlm_pred}
\title{Predicts intensities for new voxels}
\usage{
dlm_pred(newdata_df, curr_models, ptimes)
}
\arguments{
\item{newdata_df}{is a data frame with new voxel information, including variables for all terms in model
can be one or multiple voxels, but must be a dataframe}

\item{curr_models}{is a list of linear model objects, comes directly from output from dlm command (dlm_model)}

\item{ptimes}{is vector of prediction times, use for column names for new data predictions}
}
\value{
pred_obs-a matrix of prediction intensities at the times given in the input ptimes
}
\description{
Predicts intensities for new voxels
}
