% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_times.R
\name{interp_times}
\alias{interp_times}
\title{The main function takes data from MS lesions and interpolates the times to a set grid of times}
\usage{
interp_times(xdat, times, tstar)
}
\arguments{
\item{xdat}{= data_frame including variables vid, stime, y_std}

\item{times}{= times at which we interpolate the intensities}

\item{tstar}{= "current time" (divides historical times and future prediction times)
Interpolation is done linearly before tstar and flat interpolation after tstar}
}
\value{
original data (df) with the following columns/variables:
itime = interpolated time points
y_i = interpolated intensity values

The following variables must be provided:
}
\description{
The main function takes data from MS lesions and interpolates the times to a set grid of times
}
