% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dlm.R
\name{plot_dlm}
\alias{plot_dlm}
\title{Function for plotting data (predicted and original)}
\usage{
plot_dlm(dat_b, vids)
}
\arguments{
\item{dat_b}{is a data frame with the following columns/variables
lid (lesion id)
vid (voxel id)
xind,yind,zind (x,y,z coordinates for voxel location in lesion/brain)
y (voxel intensity)
source (character vector, indicating source of voxel intensity data - with two options)
"Standardized" or "Predicted"
stime (time of visits, can be interpolated times or actual times - numeric)}

\item{vids}{is one voxel id from the data (the voxel that will be plotted)}
}
\value{
plot object (grob) from grid.arrange, also outputs the plot to window
}
\description{
Needs two inputs, a data frame dat_b, and a vector of vids (voxels that are going to be highlighted)
plot_dlm is a function for just one voxel
plot_dlm2 is a function for two voxels
}
