% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dlm2.R
\name{plot_dlm2}
\alias{plot_dlm2}
\title{Function for plotting data (predicted and original)}
\usage{
plot_dlm2(dat_b, vids, aadjust = -0.5, badjust = 0.5)
}
\arguments{
\item{dat_b}{is a data frame with the following columns/variables
lid (lesion id)
vid (voxel id)
xind,yind,zind (x,y,z coordinates for voxel location in lesion/brain)
y (voxel intensity)
source (character vector, indicating source of voxel intensity data - with two options)
"Standardized" or "Predicted"
stime (time of visits, can be interpolated times or actual times - numeric)}

\item{vids}{is a vector of length 2 with voxel ids from the data (the two voxels that will be plotted)}

\item{aadjust}{is vertical adjustment for arrow label in lesion slice plot for voxel in A (first voxel)}

\item{badjust}{is vertical adjustment for arrow label in lesion slice plot for voxel in B (second voxel)}
}
\value{
plot object (grob) from grid.arrange, also outputs the plot to window
}
\description{
Needs four inputs, a data frame dat_b, and a vector of vids (voxels that are going to be highlighted)
plot_dlm is a function for just one voxel
plot_dlm2 is a function for two voxels
}
