% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_times.R
\name{time_interpolate}
\alias{time_interpolate}
\title{Interpolation functions
time_interpolate is a helper function for TimeInterpolateByGroup
that operates on each of the groups. In the input to this function,
the grouping_variable column of the data frame should be single-valued.
The function returns a (probably longer) data frame, with estimated
values for the times specified in the output_times array.}
\usage{
time_interpolate(data_frame, grouping_variable, time_var, output_times)
}
\description{
Interpolation functions
time_interpolate is a helper function for TimeInterpolateByGroup
that operates on each of the groups. In the input to this function,
the grouping_variable column of the data frame should be single-valued.
The function returns a (probably longer) data frame, with estimated
values for the times specified in the output_times array.
}
